+++ packages/cli/src/sdk/index.ts
@@ -1,0 +1,6 @@
+/**
+ * SDK module exports
+ */
+
+export * from './client';
+export * from './executor';